revised LatitudeOnLongtitude and HypocenterParser

+++ src/main/java/jp/ac/aiit/pbl/Latitude.java
@@ -1,25 +1,25 @@
 package jp.ac.aiit.pbl;
 
 public class Latitude {
-    private int degree;
-    private int minute;
-    private int second;
+    private double degree;
+    private double minute;
+    private double second;
     
-    Latitude(int degree, int minute, int second){
+    Latitude(double degree, double minute, double second){
         this.degree = degree;
         this.minute = minute;
         this.second = second;
     }
     
-    public int getDegree() {
+    public double getDegree() {
         return degree;
     }
     
-    public int getMinute() {
+    public double getMinute() {
         return minute;
     }
     
-    public int getSecond() {
+    public double getSecond() {
         return second;
     }
     
@@ -32,3 +32,4 @@ public class Latitude {
                 '}';
     }
 }
+
